accessible/tests/mochitest/name/test_nsRootAcc.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Accessibility Name Calculating Test.">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>

  <script type="application/javascript">
  <![CDATA[
    // var gA11yEventDumpID = "eventdump"; // debug stuff

    function doTest()
    {
      todo(false, "Disabled test. (Bug 586818)");
      SimpleTest.finish();
      return;

      if (LINUX) {
        todo(false, "Skip test on Linux. (Bug 525175)");
        SimpleTest.finish();
        return;
      }

      var w = window.openDialog("nsRootAcc_wnd.xul",
                                "nsRootAcc_name_test", 
                                "chrome,width=600,height=600");
    }

    SimpleTest.waitForExplicitFinish();
    addLoadEvent(doTest);
  ]]>
  </script>

  <vbox flex="1" style="overflow: auto;">
  <body xmlns="http://www.w3.org/1999/xhtml">
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=507382"
       title="focus is fired earlier than root accessible name is changed when switching between tabs">
      Mozilla Bug 
    </a>
    <p id="display"></p>
    <div id="content" style="display: none">
    </div>
    <pre id="test">
    </pre>
  </body>

  <vbox id="eventdump"></vbox>
  </vbox>
</window>
